NewPayload Tests Refactor
